+++ test_project/lib/src/domain/entities/user/user.dart
@@ -1,5 +1,0 @@
-class User {
-  final String id;
-  final String name;
-  final int age;
-}